Added new sign/symbol block ids.

+++ src/main/java/net/fexcraft/mod/famm/blocks/FAMMBlocks.java
@@ -20,7 +20,10 @@ public final class FAMMBlocks{
 					"cy_hard_sign", "cy_y", "cy_soft_sign", "cy_e", "cy_yu", "cy_ya" };
 	public static String[] latin = new String[]{ "a", "b", "c", "d", "e", "f", "g", "h", "i", "j", "k", "l", "m", "n", "o", "p", "q", "r", "s", "t", "u", "v", "w", "x", "y", "z" };
 	public static String[] numbers = new String[]{ "n0", "n1", "n2", "n3", "n4", "n5", "n6", "n7", "n8", "n9" };
-	public static String[] signs = new String[]{ "plus", "minus", "empty", "questionmark", "exclamationmark", "point", "pointdown", "comma", "colon", "space", "arrowright", "arrowleft", "arrowtop", "arrowbot" };
+	public static String[] signs = new String[]{ "plus", "minus", "empty", "questionmark", "exclamationmark", "point", "pointdown", "comma", "colon", "space",
+					"arrowright", "arrowleft", "arrowtop", "arrowbot", "et", "approx", "asterisk", "at", "lt", "gt", "lte", "gte", "degree", "slash", "backslash",
+					"equal", "hash", "infinity", "hyphen", "not_equal", "bracket_round_l", "bracket_round_r", "bracket_curly_l", "bracket_curly_r",
+					"bracket_square_l", "bracket_square_r", "bracket_angle_l", "bracket_angle_r", "percent", "promile", "pi", "sqrt" };
 	public static String[] ronum = new String[]{ "ri", "rii", "riii", "riv", "rv", "rvi", "rvii", "rviii", "rix", "rx" };
 	public static String[] deco = new String[]{ "stop", "rr", "cr", "eh", "es", "ea", "ec", "d1", "d2" };
 	public static String[] cyrillic = new String[]{ "cy_a", "cy_b", "cy_v", "cy_g", "cy_d", "cy_ie", "cy_io", "cy_zh", "cy_z", "cy_i", "cy_j", "cy_k", "cy_l",
@@ -79,6 +82,7 @@ public final class FAMMBlocks{
 	}
 
 	private static void write(String string, File file){
+		if(file.exists()) return;
 		try{
 			FileWriter writer = new FileWriter(file);
 			writer.write(string);
